+++ src/rex.mart/NEWS.rst
@@ -5,6 +5,13 @@ REX.MART Change Log
 .. contents:: Table of Contents
 
 
+0.2.0 (2016-xx-xx)
+==================
+
+- Added ability to specify additional HTSQL extensions to enable in the
+  get_mart_db() and get_mart_etl_db() functions.
+
+
 0.1.0 (2016-01-29)
 ==================